+++ packages/review-editor/hooks/tour/useTourData.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback, useRef } from 'react';
-import { DEMO_TOUR, DEMO_TOUR_ID } from '../demoTour';
+import { DEMO_TOUR, DEMO_TOUR_ID } from '../../demoTour';
 import type { CodeTourData } from '@plannotator/shared/tour';
 
 export type { TourDiffAnchor, TourKeyTakeaway, TourStop, TourQAItem, CodeTourData } from '@plannotator/shared/tour';
